+++ tui/src/network_io/protocol_types.rs
@@ -1,0 +1,22 @@
+/*
+ *   Copyright (c) 2025 R3BL LLC
+ *   All rights reserved.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+
+/// Type alias for type to read from the stream to get the length prefix.
+pub type LengthPrefixType = u64;
+/// Type aliases for the payload buffer type.
+pub type Buffer = Vec<BufferAtom>;
+pub type BufferAtom = u8;